refactor: split core models into dedicated modules
