+++ conversation_service/models/enums.py
@@ -1,0 +1,93 @@
+# Enumerations for Harena conversation service.
+from enum import Enum
+
+
+class IntentType(str, Enum):
+    """Intent types for Harena - Consultation and Analysis only."""
+
+    # === CONSULTATION (Harena Supported) ===
+    BALANCE_INQUIRY = "BALANCE_INQUIRY"
+    ACCOUNT_OVERVIEW = "ACCOUNT_OVERVIEW"
+    TRANSACTION_SEARCH = "TRANSACTION_SEARCH"
+    TRANSACTION_DETAILS = "TRANSACTION_DETAILS"
+    STATEMENT_REQUEST = "STATEMENT_REQUEST"
+
+    # === ANALYSIS (Harena Supported) ===
+    SPENDING_ANALYSIS = "SPENDING_ANALYSIS"
+    CATEGORY_ANALYSIS = "CATEGORY_ANALYSIS"
+    MERCHANT_ANALYSIS = "MERCHANT_ANALYSIS"
+    TEMPORAL_ANALYSIS = "TEMPORAL_ANALYSIS"
+    BUDGET_ANALYSIS = "BUDGET_ANALYSIS"
+    TREND_ANALYSIS = "TREND_ANALYSIS"
+    INCOME_ANALYSIS = "INCOME_ANALYSIS"
+    COMPARISON_ANALYSIS = "COMPARISON_ANALYSIS"
+
+    # === INFORMATION & SUPPORT (Harena Supported) ===
+    ACCOUNT_INFORMATION = "ACCOUNT_INFORMATION"
+    PRODUCT_INFORMATION = "PRODUCT_INFORMATION"
+    FEE_INQUIRY = "FEE_INQUIRY"
+    GENERAL_INQUIRY = "GENERAL_INQUIRY"
+    HELP_REQUEST = "HELP_REQUEST"
+
+    # === CONVERSATIONAL ===
+    GREETING = "GREETING"
+    GOODBYE = "GOODBYE"
+    THANKS = "THANKS"
+    CLARIFICATION_REQUEST = "CLARIFICATION_REQUEST"
+    POLITENESS = "POLITENESS"
+
+    # === UNSUPPORTED ACTIONS (Redirection Required) ===
+    TRANSFER_REQUEST = "TRANSFER_REQUEST"
+    PAYMENT_REQUEST = "PAYMENT_REQUEST"
+    CARD_OPERATIONS = "CARD_OPERATIONS"
+    LOAN_REQUEST = "LOAN_REQUEST"
+    ACCOUNT_MODIFICATION = "ACCOUNT_MODIFICATION"
+    INVESTMENT_OPERATIONS = "INVESTMENT_OPERATIONS"
+
+    # === MANAGEMENT CATEGORIES ===
+    UNSUPPORTED_ACTION = "UNSUPPORTED_ACTION"
+    OUT_OF_SCOPE = "OUT_OF_SCOPE"
+    UNKNOWN = "UNKNOWN"
+    AMBIGUOUS = "AMBIGUOUS"
+    INSUFFICIENT_CONTEXT = "INSUFFICIENT_CONTEXT"
+
+
+class EntityType(str, Enum):
+    """Financial entity types with hierarchical taxonomy for Harena scope."""
+
+    # === MONETARY ===
+    AMOUNT = "AMOUNT"
+    CURRENCY = "CURRENCY"
+    PERCENTAGE = "PERCENTAGE"
+
+    # === TEMPORAL ===
+    DATE_RANGE = "DATE_RANGE"
+    FREQUENCY = "FREQUENCY"
+
+    # === ACCOUNTS ===
+    ACCOUNT_TYPE = "ACCOUNT_TYPE"
+    ACCOUNT_NUMBER = "ACCOUNT_NUMBER"
+
+    # === TRANSACTIONS ===
+    MERCHANT = "MERCHANT"
+    TRANSACTION_TYPE = "TRANSACTION_TYPE"
+    CATEGORY = "CATEGORY"
+
+    # === GEOGRAPHIC ===
+    LOCATION = "LOCATION"
+    ATM_LOCATION = "ATM_LOCATION"
+
+    # === CONTACT (Often for unsupported actions) ===
+    BENEFICIARY = "BENEFICIARY"
+    COMMUNICATION = "COMMUNICATION"
+
+    # === PRODUCTS ===
+    CARD_TYPE = "CARD_TYPE"
+    LOAN_TYPE = "LOAN_TYPE"
+
+    # === SECURITY (Often for unsupported actions) ===
+    AUTHENTICATION = "AUTHENTICATION"
+    SECURITY_LEVEL = "SECURITY_LEVEL"
+
+
+__all__ = ["IntentType", "EntityType"]
